pyproject.toml
[project]
name = "fundamentos-python-puc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.41",
    "taskipy>=1.14.1",
]
[tool.taskipy.tasks]
format = """
isort .
black .
"""
